Implementada a consulta de Variação de Preço de Ativo

+++ src/VariacaoAtivo.Dominio/DTO/Quote.cs
@@ -6,7 +6,7 @@
         public float?[] low { get; set; }
         public float?[] high { get; set; }
         public float?[] close { get; set; }
-        public int?[] volume { get; set; }
+        public long?[] volume { get; set; }
     }
 
 }
